adapts fibonnaci function to take a big integer as the position instead of an integer.

+++ src/main/java/com/example/cimt/math/app/controller/MathController.java
@@ -18,7 +18,7 @@ public class MathController {
 
 
     @GetMapping("/fibonacci/{position}")
-    public ResponseEntity<BigInteger> getFibonacciNumber(@PathVariable int position) {
+    public ResponseEntity<BigInteger> getFibonacciNumber(@PathVariable BigInteger position) {
         BigInteger fiboValue = mathService.calculateFiboValue(position);
         return ResponseEntity.ok(fiboValue);
     }
